do some clean ups and refactoring
